+++ lib/nikfemm/src/magnetostatic/properties.cpp
@@ -10,7 +10,7 @@ namespace nikfemm {
         return bh_curve.size() == 0;
     }
 
-    float MagnetostaticProp::getMu(float B) const {
+    double MagnetostaticProp::getMu(double B) const {
         // mu = B / H
         // printf("B = %f\n", B);
         if (bh_curve.empty()) {
@@ -23,34 +23,34 @@ namespace nikfemm {
         BH_Curve::const_iterator it = bh_curve.lower_bound({B, 0});
         // if B is smaller than the smallest B in the curve
         if (it == bh_curve.begin()) {
-            float B1 = it->B;
-            float H1 = it->H;
-            float B2 = (++it)->B;
-            float H2 = it->H;
+            double B1 = it->B;
+            double H1 = it->H;
+            double B2 = (++it)->B;
+            double H2 = it->H;
 
             // mu = dB/dH
-            float m = (B2 - B1) / (H2 - H1);
+            double m = (B2 - B1) / (H2 - H1);
             // printf("B is smaller than the smallest B in the curve, returning mu %.17g\n", m);
             return m;
         } else if (it == bh_curve.end()) {
-            float B1 = (--it)->B;
-            float H1 = it->H;
-            float B2 = (--it)->B;
-            float H2 = it->H;
+            double B1 = (--it)->B;
+            double H1 = it->H;
+            double B2 = (--it)->B;
+            double H2 = it->H;
 
             // mu = dB/dH
-            float m = (B2 - B1) / (H2 - H1);
+            double m = (B2 - B1) / (H2 - H1);
             // printf("B is larger than the largest B in the curve, returning mu %.17g\n", m);
             return m;
         } else {
             // forward difference
-            float B1 = it->B;
-            float H1 = it->H;
-            float B2 = (++it)->B;
-            float H2 = it->H;
+            double B1 = it->B;
+            double H1 = it->H;
+            double B2 = (++it)->B;
+            double H2 = it->H;
 
             // mu = dB/dH
-            float m = (B2 - B1) / (H2 - H1);
+            double m = (B2 - B1) / (H2 - H1);
             // printf("B is in the middle of the curve, returning mu %.17g\n", m);
             return m;
         }
