+++ lib/nikfemm/src/magnetostatic/simulation.hpp
@@ -48,7 +48,7 @@ namespace nikfemm {
             Vector computeForceIntegrals(std::vector<Vector> B, SurroundingRegionBlockIntegralAssets assets, Vector p);
             double computeTorqueIntegral(std::vector<Vector> B, SurroundingRegionBlockIntegralAssets assets, Vector p, Vector center);
         protected:
-            static void updateMu(std::vector<Vector> B, std::vector<const MagnetostaticProp*>& props, std::vector<float>& mu, double residual, uint32_t iter);
+            static void updateMu(std::vector<Vector> B, std::vector<const MagnetostaticProp*>& props, std::vector<double>& mu, double residual, uint32_t iter);
             auto getSurroundingRegionBlockIntegralAssets(std::vector<Vector> B, Vector p);
     };
 }
